requirements.txt
requests==2.27.1
PyMySQL==1.0.2
PyYAML==5.3.1